requirements.txt
# Flask web framework
Flask

# Other dependencies
blinker
bcrypt
mysql-connector-python
